backend/requirements.txt
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
channels==4.0.0
click==8.1.7
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
h11==0.14.0
httptools==0.6.1
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sniffio==1.3.1
sqlparse==0.4.4
uritemplate==4.1.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
